Set the CPU model when using aarch64 architecture
Set the CPU model when using aarch64 architecture to be able to correctly
emulate the architecture from a x86_64 platform. KVM virtualization solution
is chosen when you are running a VM using the host architecture and QEMU is
used when you are emulating another architecture than the one used by the host.

Change-Id: Icd58fad778961b99ab5a77a699ed1522e2c1ed38
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>